00-bootcamp-project/scraping/pyproject.toml
[tool.poetry]
name = "scraping"
version = "0.1.0"
description = ""
authors = ["zkan <zkan.cs@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Scrapy = "^2.8.0"
google-cloud-storage = "^2.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
